requirements.txt
# Investment Dashboard Dependencies

# Stock market data
yfinance>=0.2.18

# RSS feed parsing
feedparser==6.0.10

# HTTP requests
requests==2.32.0

# Date parsing utilities
python-dateutil==2.8.2
